feat(ci): add bump chart
