+++ examples/codesandbox/src/App.tsx
@@ -1,4 +1,4 @@
-import { AreaBump, Bar, Bullet } from './charts'
+import { AreaBump, Bar, Bullet, Bump } from './charts'
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
 import ChartContainer from './ChartContainer'
 import Navigation from './Navigation'
@@ -22,6 +22,7 @@ export default function App() {
               <Route path="area-bump" element={<AreaBump />} />
               <Route path="bar" element={<Bar />} />
               <Route path="bullet" element={<Bullet />} />
+              <Route path="bump" element={<Bump />} />
             </Routes>
           </ChartContainer>
         </div>
