+++ examples/codesandbox/src/charts/index.ts
@@ -1,3 +1,3 @@
-export * from './AreaBump'
 export * from './Bar'
 export * from './Bullet'
+export * from './Bump'
